pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "ipython>=8.31.0",
    "isort>=5.13.2",
    "langchain>=0.3.13",
    "langchain-ollama>=0.2.2",
    "langchainhub>=0.1.21",
    "mypy>=1.14.0",
    "python-dotenv>=1.0.1",
    "requests==2.27.1",
    "ruff>=0.8.4",
]
